Minor code cleanup. MigrationUtils should be moved and all its methods made protected (as they bypass Hibernate). Update description of 'repair' command in DatabaseUtils.
